get user data with token (core) create settings page
